Restructure the file hierarchy
We have so many notebooks we need a directory for the notebooks.
Again, this will probably break a lot of things. We shall fix issues as they appear.

Added a 32 MB catalog from GLEAM which includes the 'alpha' column. If you do not end up needing this, you can always adjust the parser so that it ignores the last item.