Split the file hierarchy; add notebook for testing
There is a new notebook directory called visibility, which shall house the materials on which I have been recently working. The point is to allow relative imports for the new single_source notebook, which is intended to test the calculations I have coded. Now I have to switch to Askia to perform actual testing, because Windows is generally incompatible with astro research :(