First steps toward creating an importable package
These scripts augment the notebooks, but I find that I am frequently repeating myself. I want to centralize them and I want to be able to import them from anywhere.

I am starting a new branch because I am entirely new to package development. This could all turn out to be a wild goose chase, in which case I will have to go back to copying and pasting the scripts like a fool.