FIxed 2 disastrous mistakes
1. The J_matrix function was setting the LST to the current LST regardless of what the user actually input.

2. The get_lst function was using the HERA longitude regardless of what the user actually input.

This commit includes the quality of life improvement of making get_lst more consistent. Specifically, the old form involved degrees as input and radians as output. The new form involves 'radians' as a paramater, which sets the format for both the input and the output.

WARNING: this is NOT a stable version! I have to correct EVERY reference to get_lst to fit the new unit standard!