Repaired and partially upgraded beam_response_gaussian
The results do not look quite as they used to. The code may have become more accurate, or the additional layers of complexity might have allowed an error to sneak in somewhere. Hard to say.

On the bright side, the shape of the "Gaussian" has changed somewhat. I think it is less correct than it was before, but it is at least something I can bring up to PI, to verify.