At last
We have run some conclusive tests, and we have a better picture as to the problem.

Interpolation is going to be processing a data set at least a dozen thousand long, multiplied by the number of matrix entries we want to fetch.

The next idea I have, re: optimizing the sky-over-LST query, is to find a closed-form solution for rotating the Jones matrix. Or even the A matrix, if I can expand it out.