Keeping track of appearances
This is important, because the demo.py script is designed for user-friendliness. If we have no notebook featuring the function, we either need to create one, or we need to scrap the function.

I am bumping up the version due to the recent expansion of demo.py, but I would like to eventually split up the demo.py script. We should have a folder called 'demos' and there should be a separate script for each skyflux script that we are preparing demos for...

In any case, we will want to find *some* way of splitting up demos.py, otherwise it will grow intolerably long, I project.