Breaking extrapolated frequencies again
I remembered that my objection was not merely philosophical. My linear interpolator would require a few extra bits of logic to even be able to interpret such requests. Reverting from a warning system to a full breaking system.

This also means that I will be manually restricting the scope of the visibility tensor experiment. We will stay safely in the 77-226 MHz range.