Finished backing up the visibility tensor
I think I am ready for a trial run, as soon as the installation finishes. We expect crashes galore, but I am merging this branch back into master anyway, because this is technically a stable commit (cold_tensor is an optional script) and because we accomplished our namesake a long time ago.