Closed the loop
I forgot to add a statement which writes the final results. Lucky to catch that by eye!

Also, I included another datum in the npz save. "Dying index" is a one-of-a-kind array that simply lets me know the last source index that we saved. Therefore, if the environment crashed so badly that I even lost the shell, I should still be able to recover.

No safety, of course, if the hard drive becomes inaccessible.