Deploying half-pre-baked spline-beam-func
I promise that the messy file hierarchy forced my hand. I appreciate that there are more important things for me to be doing than goofing around with package maintenance. But in this case I virtually *needed* to sink a couple of hours into it because my broken spline beam function finally caught up to me.